Abstractions to compile systemlib for Linux and Nuttx
Modified uint32_t casts of pointers to unsigned long for portability.
It otherwise breaks on x86_64.

Added _PX4_IOC to handle the conflice between _IOC on Linux and NuttX.

Removed use of px4::ok() because it cannot be used in a thread based
implementation. Changed to use px4::AppMgr which uses ok() on ROS.

Removed up_cxxinitialize.c from Linux build.

+++ src/modules/systemlib/otp.c
@@ -170,7 +170,7 @@ void F_lock(void)
 }
 
 // flash write word.
-int F_write_word(uint32_t Address, uint32_t Data)
+int F_write_word(unsigned long Address, uint32_t Data)
 {
 	unsigned char octet[4] = {0, 0, 0, 0};
 
@@ -185,7 +185,7 @@ int F_write_word(uint32_t Address, uint32_t Data)
 }
 
 // flash write byte
-int F_write_byte(uint32_t Address, uint8_t Data)
+int F_write_byte(unsigned long Address, uint8_t Data)
 {
 	volatile int status = F_COMPLETE;
